requirements.txt
alembic==0.8.3
algoliasearch==1.7.1
aniso8601==1.1.0
blinker==1.4
boto3==1.2.2
botocore==1.3.9
docutils==0.12
factory-boy==2.6.0
fake-factory==0.5.3
Flask==0.10.1
Flask-Login==0.3.2
Flask-Migrate==1.6.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
hashids==1.1.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
Mako==1.0.3
MarkupSafe==0.23
marshmallow==2.3.0
newrelic==2.58.2.45
psycopg2==2.6.1
python-dateutil==2.4.2
python-editor==0.4
python-magic==0.4.10
pytz==2015.7
requests==2.8.1
six==1.10.0
SQLAlchemy==1.0.9
Werkzeug==0.11.2
